refactor: extract GoodsMap and share between Inventory and PriceRanges
